Initial commit, pre-release of fdaPDE module
Initial branch commit. Pre-release of bathymetry module implementing fdaPDE process developed by Wang, R. & Ateljevich, E., Fregoso, T. A. & Jaffe, B. E. (2018), A Revised Continuous Surface Elevation Model for Modeling (Chapter 5). In Methodology for Flow and Salinity Estimates in the Sacramento-San Joaquin Delta and Suisun Marsh, 38th Annual Progress Report to the State Water Resources Control Board. Module uses arcpy to partially automate steps of interpolating bathymetry points to generate a bathymetry surface and merge bathymetry surface with LiDAR.

This is a pre-release version of this module and is currently under development, testing, and documentation. All variable names and functionality are subject to change. THIS MODULE IS PROVIDED ON AN "AS-IS" BASIS ONLY.